requirements.txt
discord.py
tinydb
git+https://github.com/LIAAD/yake